fix: 收款课时数parseInt截断小数(6.6→6)改为parseFloat
25分钟课10节=6.6课时, parseInt(6.6)=6 截断小数
改为 parseFloat(6.6)=6.6 正确保存小数课时

+++ src/pages/Payments.jsx
@@ -68,7 +68,7 @@ export default function Payments() {
         date: formData.date,
         payment_method: formData.method,  // 修正字段名
         description: formData.notes,
-      hours: parseInt(formData.packageHours) || 0
+      hours: parseFloat(formData.packageHours) || 0
       });
 
       // addStudentHours removed: backend already updates total_hours in payments.js
